+++ src/Phan/Language/Element/Comment/Builder.php
@@ -76,7 +76,6 @@ final class Builder
         int $lineno,
         int $comment_type
     ) {
-        // @phan-suppress-next-line PhanPossiblyFalseTypeMismatchProperty
         $this->lines = \explode("\n", $comment);
         $this->comment_lines_count = \count($this->lines);
         $this->code_base = $code_base;
@@ -296,7 +295,7 @@ final class Builder
             $this->magic_property_list,
             $this->magic_method_list,
             $this->phan_overrides,
-            $this->closure_scope,  // @phan-suppress-current-line PhanPartialTypeMismatchArgument
+            $this->closure_scope,
             $this->throw_union_type,
             // NOTE: The code base and context are used for emitting issues, and are not saved
             $this->code_base,
@@ -311,7 +310,7 @@ final class Builder
     {
         // https://secure.php.net/manual/en/regexp.reference.internal-options.php
         // (?i) makes this case sensitive, (?-1) makes it case insensitive
-        if (\preg_match('/@((?i)param|var|return|throws|throw|returns|inherits|suppress|phan-[a-z0-9_-]*(?-i)|method|property|template|PhanClosureScope)\b/', $line, $matches)) {
+        if (\preg_match('/@((?i)param|var|return|throws|throw|returns|inherits|suppress|phan-[a-z0-9_-]*(?-i)|method|property|property-read|property-write|template|PhanClosureScope)(?:[^a-zA-Z0-9_\x7f-\xff-]|$)/', $line, $matches)) {
             $case_sensitive_type = $matches[1];
             $type = \strtolower($case_sensitive_type);
 
@@ -343,7 +342,7 @@ final class Builder
                 );
             } elseif ($type === 'suppress') {
                 $this->maybeParseSuppress($i, $line);
-            } elseif ($type === 'property') {
+            } elseif ($type === 'property' || $type === 'property-read' || $type === 'property-write') {
                 $this->maybeParseProperty($i, $line);
             } elseif ($type === 'method') {
                 $this->maybeParseMethod($i, $line);
@@ -695,8 +694,8 @@ final class Builder
     private static function templateTypeFromCommentLine(
         string $line
     ) {
-        // TODO: Just use WORD_REGEX? Backslashes or nested templates wouldn't make sense.
-        if (preg_match('/@(?:phan-)?template\s+(' . Type::simple_type_regex . ')/', $line, $match)) {
+        // Backslashes or nested templates wouldn't make sense, so use WORD_REGEX.
+        if (preg_match('/@(?:phan-)?template\s+(' . self::WORD_REGEX . ')/', $line, $match)) {
             $template_type_identifier = $match[1];
             return new TemplateType($template_type_identifier);
         }
@@ -905,8 +904,6 @@ final class Builder
      *
      * @return Property|null
      * magic property with the union type.
-     *
-     * TODO: guess line number for emitted issue
      */
     private function magicPropertyFromCommentLine(
         string $line,
@@ -914,7 +911,6 @@ final class Builder
     ) {
         // Note that the type of a property can be left out (@property $myVar) - This is equivalent to @property mixed $myVar
         // TODO: properly handle duplicates...
-        // TODO: support read-only/write-only checks elsewhere in the codebase?
         if (\preg_match('/@(?:phan-)?(property|property-read|property-write)(?:\s+(' . UnionType::union_type_regex . '))?(?:\s+(?:\\$' . self::WORD_REGEX . '))/', $line, $match)) {
             $category = $match[1];
             if ($category === 'property-read') {