+++ src/Phan/Language/Type/LiteralStringType.php
@@ -4,6 +4,7 @@ namespace Phan\Language\Type;
 use InvalidArgumentException;
 use Phan\Config;
 use Phan\Language\Type;
+use Phan\Language\UnionType;
 use RuntimeException;
 
 /**
@@ -127,8 +128,10 @@ final class LiteralStringType extends StringType implements LiteralTypeInterface
                     return self::UNESCAPE_CHARACTER_LOOKUP[$x];
                 }
                 // convert 2 hex bytes to a single character
+                // @phan-suppress-next-line PhanPossiblyFalseTypeArgumentInternal
                 return \chr(\hexdec(\substr($x, 2)));
             },
+            // @phan-suppress-next-line PhanPossiblyFalseTypeArgumentInternal
             $escaped_string
         );
         return self::instanceForValue($escaped_string, $is_nullable);
@@ -235,6 +238,13 @@ final class LiteralStringType extends StringType implements LiteralTypeInterface
     public function canSatisfyComparison($scalar, int $flags) : bool
     {
         return self::performComparison($this->value, $scalar, $flags);
+    }
+
+    public function getTypeAfterIncOrDec() : UnionType
+    {
+        $v = $this->value;
+        ++$v;
+        return Type::nonLiteralFromObject($v)->asUnionType();
     }
 }