+++ NEWS.md
@@ -1,18 +1,30 @@
 Phan NEWS
 
-?? ??? 2018, Phan 1.1.3 (dev)
+?? ??? 2018, Phan 1.1.4 (dev)
 -----------------------
 
-New features(CLI)
+Maintenance:
++ Emit `UnusedPluginSuppression` on `@phan-suppress-next-line` and `@phan-file-suppress`
+  on the same line as the comment declaring the suppression. (#2167, #1731)
++ Don't emit `PhanInvalidCommentForDeclarationType` (or attempt to parse) unknown tags that have known tags as prefixes  (#2156)
+  (e.g. `@param-some-unknown-tag`)
+
+Bug fixes:
++ Fix a crash when analyzing a nullable parameter of type `self` in traits (#2163)
++ Properly parse closures/generic arrays/array shapes when inner types also contain commas (#2141)
++ Support matching parentheses inside closure params, recursively. (e.g. `Closure(int[],Closure(int):bool):int[]`)
++ Don't warn about properties being read-only when they might be modified by reference (#1729)
+
+20 Nov 2018, Phan 1.1.3
+-----------------------
+
+New features(CLI):
 + Warn when calling method on union types that are definitely partially invalid. (#1885)
   New config setting: `--strict-method-checking` (enabled as part of `--strict-type-checking`)
   New issue type: `PhanPossiblyNonClassMethodCall`
 + Add a prototype tool `tool/phoogle`, which can be used to search for function/method signatures in user-declared and internal functions/methods.
   E.g. to look for functions that return a string, given a string and an array:
   `/path/phan/tool/phoogle 'string -> array -> string`
-
-Plugins:
-+ Add `BeforeAnalyzeCapability`, which will be executed once before starting the analysis phase. (#2086)
 
 New features(Analysis):
 + Add a heuristic check to detect potential infinite recursion in a functionlike calling itself (i.e. stack overflows)
@@ -30,11 +42,18 @@ New features(Analysis):
   New issue types: `PhanTypeInvalidCallable`, `PhanTypePossiblyInvalidCallable` (the latter check requires `--strict-method-checking`)
 + Reduce false positives for a few functions (such as `substr`) in strict mode.
 + Make Phan infer that variables are not null/false from various comparison expressions, e.g. `assert($x > 0);`
++ Detect invalid arguments to `++`/`--` operators (#680).
+  Improve the analysis of the side effects of `++`/`--` operators.
+  New issue type: `PhanTypeInvalidUnaryOperandIncOrDec`
+
+Plugins:
++ Add `BeforeAnalyzeCapability`, which will be executed once before starting the analysis phase. (#2086)
 
 Bug fixes:
 + Fix false positives analyzing `define()` (#2128)
 + Support declaring instance properties as the union type `static` (#2145)
   New issue types: `PhanStaticPropIsStaticType`
++ Fix a crash seen when Phan attempted to emit `PhanTypeArrayOperator` for certain operations (#2153)
 
 05 Nov 2018, Phan 1.1.2
 -----------------------